Add remaining missing record DB functionality to cf.storage.tau_enterprise and add tests for same.
Replace AbstractStorage.match with search_inside, an operation which
checks whether a field is either equal to a value, or is a collection
which contains that value, and replace the one usage of match with
search_inside. This is necessary because the functionality of match is
difficult to implement in a backend-agnostic manner.