Specify a database name when using REST API.
The server now looks for a header 'Database-Name' specifying which
database the REST API should be exposing for this request.
This is now specified in the client in connect_database().
The test suite uses a random database name so concurrently-running tests
won't clobber one another.

This will integrate with user authentication once that is implemented.