Fixes for wrong type in StorageRecord (needs to be storage class, not database) and non-integer eids.
With these changes and the latest server version, TauEnterpriseStorage
is capable of being used wherever LocalFileStorage is currently used.
However, performance is currently poor is latency to the server is high
(e.g., server on east coast, client on west coast) due to a large number
of individual database operations. The controller will need to be
refactored to perform fewer queries.