Client support for server-side foreign reference population.
This is the first in a series of changes to move field population
(embedding copies of foreign references in the original object) and
association/disassociation (propagating back-references to foreign keys)
from the controller to the server.

The server now supports server-side field embedding. GET operations can
include the optional field ?embedded=X, where X is a list indicating
which fields should be populated in the result. Fields which are foreign
keys or lists of foreign keys may be populated server-side. This will
allow the controller's populate method on a list of N items to complete
in one network request instead of N requests.