Initial work towards client support for server-side propagation. Insert propagation for remote databases is now supported. Delete propagation is not implemented as it is not yet working server-side.
The server now supports an additional parameter ?propagate=true. If
enabled, propagation will handle on the server the "associate" operation
currently performed by the client's Controller. For each added foreign
key reference in a field declarating a backreference (a 'via' in the
model), the server will update the corresponding foreign key with a
backreference. This reduces the number of network operations required to
maintain associations.