Generalize commit bcd8a58 for BlueGene and Cray KNL.
It may be that more than one family does not have a family regex,
so assuming !family_regex implies the family must be 'System'
is not always true.

Instead, check the command list for all families that don't have a
regex and use that as a possible match.  Regex is used first since
command names are unreliable.