Logic for deciding what to push is implemented.
To enable partial push, the following logic is used:
Edges in the object graph are given a direction, 'up' and 'down'.
When pushing a specific object in the graph, we also must push those
objects which are transitively reachable along the 'down' edges
considering both 'up' and 'down' edges, but along the 'up' edges, only
those objects which are trasnsitively reachable along 'up edges'. Thus,
if we push an Experiment, we push all the child Trials with it (they are
reachable through 'down' edges), but if we push a specific Trial, we
push the Trial's experiment (along an up edge) but NOT other Trials in
the same experiment (because once we follow an up edge, we do not
subsequently follow a down edge)