Upgrade TinyDB from 2.2.1 to 3.5.0
TinyDB prior to version 3.0 had a bug where the LRU part of the LRU
cache was broken: the __eq__ function of Query objects mutated the
Query, so checking whether a Query was already in the queue changed the
query. With certain combinations of queries, this would result in a
KeyError inside TinyDB, causing TAU Commander to abort.