Demote role mismatch errors to warnings.
When discovering all installed compilers in a family, if a compiler
wrapper wraps a compiler which cannot perform the same role as the
wrapper, then warn the user and don't include that compiler wrapper
in the family.

Example: Intel C compiler doesn't support UPC.  On Cray, `cc` compiles
both C and UPC, so with PrgEnv-Intel we'll get a UPC compiler wrapper
(i.e. `cc`) which wrapps a non-UPC compiler (i.e. `icc`).  That's OK.
Cray expects the user to know the difference between C and UPC and to
not compile UPC with PrgEnv-Intel loaded.  In this case, TAU Commander
will act as if no UPC compiler is available and warn the user.