Work towards visualization and analysis support.
`AbstractAnalysis` class defines an analysis. An analysis can either be
run directly (e.g., interactively in a Jupyter notebook), or an analysis
can create Jupyter InputCells and write those cells to a new notebook
for use.

A simple horizontal bar chart example is provided as an
example. Selecting a Trial in the JupyterHub dashboard creates a new
notebook containing code to generate the chart, then opens and
executes that notebook.