Click on axis label in trial barplot to view profile barplot.
As in ParaProf, clicking the axis label in the bar plot for a trial will
now open a view showing the plot for that specific (node, context,
thread) profile. This is done by having a Bokeh JavaScript callback
invoke the TAU Commander JupyterLab extension, which in turn
sends a message to the Python backend to create a new analysis notebook
prepopulated with the requsted (n,c,t) setting and opening it.