Split 'run' and 'show' in CorrelationAnalysis
The analysis can be used both to produce a chart for viewing in the
notebook interface or for getting a correlation coefficient. Displaying
the chart is now separated from the data analysis that goes into it, as
attempting to display a chart outside the notebook interface changes
output parameters in a way that can break subsequent code. This fixes
test failures after test_correlation_against_self