Renumber entire list of trial numbers at once.
Previous implementation of `trial renumber` subcommand did a pair-wise
swap of trial numbers. This commit allows for renumbering an entire
list of trial numbers at once.

Also checks if a renumbering will delete a trial and tells the user
to use the `trial delete` subcommand to delete a trial rather than
using this subcommand.