Fix install error with IBM CUDA Fortran compiler
When the using the IBM family compilers, a CUDA Fortran compiler is provided (xlcuf), but not a CUDA C++ compiler. This caused an error when trying to select the default compiler for the CUDA_CXX role.

This change checks whether the family actually supports a given role before trying to use it as the default.

Fixes #347.