Fix for exception on Docker Fuse w/ shutil.move
After upgrading to python3, shutil.move() fails sometimes with symlinks on fuse
mounted file systems. The copies appear to happen without issue, but
calls to copystat() from within copytree() raise an error. If shutil.move()
raises an error, but the files in the error message are in the destination
attempt to ignore it and continue.