Use `INSTALLDIR` over prefix in setup.cfg
Same for the build directory.
The makefile was replacing `INSTALLDIR` even if it's empty in
`setup.cfg`
User set environment variables should take precidence over (possibly)
edited config files.
It might be better to inspect the output of the shell command and only
assign when a value exists but the logic of when the environment
overrides the config file starts to get complicated.